登录功能完成

+++ src/main/java/com/sweetpotatoclock/service/impl/GoalCompleteServiceImpl.java
@@ -7,6 +7,7 @@ import com.sweetpotatoclock.service.GoalCompleteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
@@ -64,7 +65,7 @@ public class GoalCompleteServiceImpl implements GoalCompleteService {
         GoalComplete goalComplete=new GoalComplete();
         goalComplete.setGroupId(groupId);
         goalComplete.setUserId(userId);
-        goalComplete.setCompletion(0);
+        goalComplete.setCompletion(BigDecimal.valueOf(0));
         goalComplete.setIsClocked(0);
         if(goalCompleteMapper.insert(goalComplete)==1){
             return true;
